2444. Count Subarrays With Fixed Bounds.
